+++ typescript/aws-codepipeline-ecs-lambda/lib/pipeline-app-async-lambda-stack.ts
