+++ typescript/aws-codepipeline-ecs-lambda/lib/pipeline-app-lambda-api-stack.ts
