+++ typescript/aws-codepipeline-ecs-lambda/lib/pipeline-stack.ts
@@ -1,16 +1,15 @@
 import * as cdk from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import { CodePipeline, CodePipelineSource, ManualApprovalStep, ShellStep, Wave } from 'aws-cdk-lib/pipelines';
-import { pipelineAppStage } from './pipeline-app-stage';
+import { pipelineAppStage } from './app-stage';
 
 export class pipelineStack extends cdk.Stack {
   constructor(scope: Construct, id: string, props?: cdk.StackProps) {
     super(scope, id, props);
 
-    const githubOrg       = process.env.GITHUB_ORG       || "jfan9";
-    const githubRepo      = process.env.GITHUB_REPO      || "jfan9-aws-cdk-examples";
-    const githubProject   = process.env.GITHUB_PROJECT   || "typescript/aws-codepipeline-ecs-lambda";
-    const githubBranch    = process.env.GITHUB_BRANCH    || "issue#820";
+    const githubOrg       = process.env.GITHUB_ORG       || "aws-6w8hnx";
+    const githubRepo      = process.env.GITHUB_REPO      || "aws-codepipeline-ecs-lambda";
+    const githubBranch    = process.env.GITHUB_BRANCH    || "main";
     const devEnv          = process.env.DEV_ENV          || "dev";
 
     const pipeline = new CodePipeline(this, 'pipeline', {
@@ -18,7 +17,7 @@ export class pipelineStack extends cdk.Stack {
       crossAccountKeys: true,
       reuseCrossRegionSupportStacks: true,
       synth: new ShellStep('Synth', {
-        input: CodePipelineSource.connection(`${githubOrg}/${githubRepo}/${githubProject}`, `${githubBranch}`,{
+        input: CodePipelineSource.connection(`${githubOrg}/${githubRepo}`, `${githubBranch}`,{
           // You need to replace the below code connection arn:
           connectionArn: `arn:aws:codestar-connections:ap-southeast-2:${props?.env?.account}:connection/0ce75950-a29b-4ee4-a9d3-b0bad3b2c0a6`
         }),
